+++ examples/data-sources/cpanel_domains/data-source.tf
@@ -1,0 +1,1 @@
+data "cpanel_domains" "all" {}